add buy cli
